Revert "Small changes MAC"
This reverts commit 8c9cc1979e43c7dbd154e483bad91c5871ba3bec.

+++ src/NPC.cpp
@@ -81,8 +81,7 @@ int NPC::GetLocation(UWorldTime &t)
 				diff = 0;
 			if (diff >= it->second->_nodes.size())
 				diff = it->second->_nodes.size() - 1;
-			position = (intptr_t)it->second->_nodes.at(diff);
-            //position = (int)it->second->_nodes.at(diff);
+			position = (int)it->second->_nodes.at(diff);
 		}
 	}
 	return position;
